Version: 0.1.10 - Updated UserControl1.  Added functionality to Add Customer and Clear buttons.

+++ COMP212-Midterm-SalePOS/ManagementForm.cs
@@ -34,7 +34,7 @@ namespace COMP212_Midterm_SalePOS
         {
             if (NoEmptyFields())
             {
-                Connection.AddRow(int.Parse(IDTextBox.Text), ProductNameTextBox.Text, CategoryComboBox.SelectedItem.ToString(), int.Parse(QuantityTextBox.Text), int.Parse(PriceTextBox.Text), int.Parse(SalePriceTextBox.Text), UserTextBox.Text, DescriptionTextBox.Text);
+                Connection.AddRow(int.Parse(IDTextBox.Text), ProductNameTextBox.Text, CategoryComboBox.SelectedItem.ToString(), int.Parse(QuantityTextBox.Text), double.Parse(PriceTextBox.Text), double.Parse(SalePriceTextBox.Text), UserTextBox.Text, DescriptionTextBox.Text);
             }
             else
             {
